Add better support for large files
This uses the old trick of setting the Response entity to a StreamingOutput instance, instead of a byte array.

To make this work with the CacheHandler, I had to add access to set a supplier for a tee stream to the StreamingOutput interface, which is how we allow caching efficiently. This means the CacheHandler will register its secondary stream as a Supplier<OutputStream> on the StreamingOutput.

NOTE: I haven't verified that this actually reduces memory usage. It's still possible that Quarkus doesn't handle the conversion from StreamingOutput to wire protocol efficiently, and we could end up with the same memory bloat in the Quarkus libraries. We need to verify this using JVM monitoring during a test.